+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/SessionPoolOptions.java
@@ -613,19 +613,19 @@ public class SessionPoolOptions {
 
     // This field controls the default behavior of session management in Java client.
     // Set useMultiplexedSession to true to make multiplexed session the default.
-    private boolean useMultiplexedSession = false;
+    private boolean useMultiplexedSession = true;
 
     // This field controls the default behavior of session management for RW operations in Java
     // client.
     // Set useMultiplexedSessionForRW to true to make multiplexed session for RW operations the
     // default.
-    private boolean useMultiplexedSessionForRW = false;
+    private boolean useMultiplexedSessionForRW = true;
 
     // This field controls the default behavior of session management for Partitioned operations in
     // Java client.
     // Set useMultiplexedSessionPartitionedOps to true to make multiplexed session for Partitioned
     // operations the default.
-    private boolean useMultiplexedSessionPartitionedOps = false;
+    private boolean useMultiplexedSessionPartitionedOps = true;
 
     private Duration multiplexedSessionMaintenanceDuration = Duration.ofDays(7);
     private Clock poolMaintainerClock = Clock.INSTANCE;