+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/SpannerCloudMonitoringExporterUtils.java
@@ -22,6 +22,7 @@ import static com.google.api.MetricDescriptor.MetricKind.UNRECOGNIZED;
 import static com.google.api.MetricDescriptor.ValueType.DISTRIBUTION;
 import static com.google.api.MetricDescriptor.ValueType.DOUBLE;
 import static com.google.api.MetricDescriptor.ValueType.INT64;
+import static com.google.cloud.spanner.BuiltInMetricsConstant.ALLOWED_EXEMPLARS_ATTRIBUTES;
 import static com.google.cloud.spanner.BuiltInMetricsConstant.GAX_METER_NAME;
 import static com.google.cloud.spanner.BuiltInMetricsConstant.GRPC_METER_NAME;
 import static com.google.cloud.spanner.BuiltInMetricsConstant.PROJECT_ID_KEY;
@@ -36,17 +37,24 @@ import com.google.api.Metric;
 import com.google.api.MetricDescriptor.MetricKind;
 import com.google.api.MetricDescriptor.ValueType;
 import com.google.api.MonitoredResource;
+import com.google.monitoring.v3.DroppedLabels;
 import com.google.monitoring.v3.Point;
+import com.google.monitoring.v3.SpanContext;
 import com.google.monitoring.v3.TimeInterval;
 import com.google.monitoring.v3.TimeSeries;
 import com.google.monitoring.v3.TypedValue;
+import com.google.protobuf.Any;
+import com.google.protobuf.Timestamp;
 import com.google.protobuf.util.Timestamps;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.sdk.metrics.data.AggregationTemporality;
+import io.opentelemetry.sdk.metrics.data.DoubleExemplarData;
 import io.opentelemetry.sdk.metrics.data.DoublePointData;
+import io.opentelemetry.sdk.metrics.data.ExemplarData;
 import io.opentelemetry.sdk.metrics.data.HistogramData;
 import io.opentelemetry.sdk.metrics.data.HistogramPointData;
+import io.opentelemetry.sdk.metrics.data.LongExemplarData;
 import io.opentelemetry.sdk.metrics.data.LongPointData;
 import io.opentelemetry.sdk.metrics.data.MetricData;
 import io.opentelemetry.sdk.metrics.data.MetricDataType;
@@ -57,6 +65,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import java.util.stream.Collectors;
 
 class SpannerCloudMonitoringExporterUtils {
 
@@ -69,7 +78,8 @@ class SpannerCloudMonitoringExporterUtils {
     return resource.getAttributes().get(PROJECT_ID_KEY);
   }
 
-  static List<TimeSeries> convertToSpannerTimeSeries(List<MetricData> collection) {
+  static List<TimeSeries> convertToSpannerTimeSeries(
+      List<MetricData> collection, String projectId) {
     List<TimeSeries> allTimeSeries = new ArrayList<>();
 
     for (MetricData metricData : collection) {
@@ -94,7 +104,8 @@ class SpannerCloudMonitoringExporterUtils {
       metricData.getData().getPoints().stream()
           .map(
               pointData ->
-                  convertPointToSpannerTimeSeries(metricData, pointData, monitoredResourceBuilder))
+                  convertPointToSpannerTimeSeries(
+                      metricData, pointData, monitoredResourceBuilder, projectId))
           .forEach(allTimeSeries::add);
     }
     return allTimeSeries;
@@ -103,7 +114,8 @@ class SpannerCloudMonitoringExporterUtils {
   private static TimeSeries convertPointToSpannerTimeSeries(
       MetricData metricData,
       PointData pointData,
-      MonitoredResource.Builder monitoredResourceBuilder) {
+      MonitoredResource.Builder monitoredResourceBuilder,
+      String projectId) {
     TimeSeries.Builder builder =
         TimeSeries.newBuilder()
             .setMetricKind(convertMetricKind(metricData))
@@ -135,7 +147,7 @@ class SpannerCloudMonitoringExporterUtils {
             .setEndTime(Timestamps.fromNanos(pointData.getEpochNanos()))
             .build();
 
-    builder.addPoints(createPoint(metricData.getType(), pointData, timeInterval));
+    builder.addPoints(createPoint(metricData.getType(), pointData, timeInterval, projectId));
 
     return builder.build();
   }
@@ -191,7 +203,7 @@ class SpannerCloudMonitoringExporterUtils {
   }
 
   private static Point createPoint(
-      MetricDataType type, PointData pointData, TimeInterval timeInterval) {
+      MetricDataType type, PointData pointData, TimeInterval timeInterval, String projectId) {
     Point.Builder builder = Point.newBuilder().setInterval(timeInterval);
     switch (type) {
       case HISTOGRAM:
@@ -199,7 +211,8 @@ class SpannerCloudMonitoringExporterUtils {
         return builder
             .setValue(
                 TypedValue.newBuilder()
-                    .setDistributionValue(convertHistogramData((HistogramPointData) pointData))
+                    .setDistributionValue(
+                        convertHistogramData((HistogramPointData) pointData, projectId))
                     .build())
             .build();
       case DOUBLE_GAUGE:
@@ -221,7 +234,7 @@ class SpannerCloudMonitoringExporterUtils {
     }
   }
 
-  private static Distribution convertHistogramData(HistogramPointData pointData) {
+  private static Distribution convertHistogramData(HistogramPointData pointData, String projectId) {
     return Distribution.newBuilder()
         .setCount(pointData.getCount())
         .setMean(pointData.getCount() == 0L ? 0.0D : pointData.getSum() / pointData.getCount())
@@ -229,6 +242,71 @@ class SpannerCloudMonitoringExporterUtils {
             BucketOptions.newBuilder()
                 .setExplicitBuckets(Explicit.newBuilder().addAllBounds(pointData.getBoundaries())))
         .addAllBucketCounts(pointData.getCounts())
+        .addAllExemplars(
+            pointData.getExemplars().stream()
+                .map(e -> mapExemplar(e, projectId))
+                .collect(Collectors.toList()))
         .build();
+  }
+
+  private static Distribution.Exemplar mapExemplar(ExemplarData exemplar, String projectId) {
+    double value = 0;
+    if (exemplar instanceof DoubleExemplarData) {
+      value = ((DoubleExemplarData) exemplar).getValue();
+    } else if (exemplar instanceof LongExemplarData) {
+      value = ((LongExemplarData) exemplar).getValue();
+    }
+
+    Distribution.Exemplar.Builder exemplarBuilder =
+        Distribution.Exemplar.newBuilder()
+            .setValue(value)
+            .setTimestamp(mapTimestamp(exemplar.getEpochNanos()));
+    if (exemplar.getSpanContext().isValid()) {
+      exemplarBuilder.addAttachments(
+          Any.pack(
+              SpanContext.newBuilder()
+                  .setSpanName(
+                      makeSpanName(
+                          projectId,
+                          exemplar.getSpanContext().getTraceId(),
+                          exemplar.getSpanContext().getSpanId()))
+                  .build()));
+    }
+    if (!exemplar.getFilteredAttributes().isEmpty()) {
+      exemplarBuilder.addAttachments(
+          Any.pack(mapFilteredAttributes(exemplar.getFilteredAttributes())));
+    }
+    return exemplarBuilder.build();
+  }
+
+  static final long NANO_PER_SECOND = (long) 1e9;
+
+  private static Timestamp mapTimestamp(long epochNanos) {
+    return Timestamp.newBuilder()
+        .setSeconds(epochNanos / NANO_PER_SECOND)
+        .setNanos((int) (epochNanos % NANO_PER_SECOND))
+        .build();
+  }
+
+  private static String makeSpanName(String projectId, String traceId, String spanId) {
+    return String.format("projects/%s/traces/%s/spans/%s", projectId, traceId, spanId);
+  }
+
+  private static DroppedLabels mapFilteredAttributes(Attributes attributes) {
+    DroppedLabels.Builder labels = DroppedLabels.newBuilder();
+    attributes.forEach(
+        (k, v) -> {
+          String key = cleanAttributeKey(k.getKey());
+          if (ALLOWED_EXEMPLARS_ATTRIBUTES.contains(key)) {
+            labels.putLabel(key, v.toString());
+          }
+        });
+    return labels.build();
+  }
+
+  private static String cleanAttributeKey(String key) {
+    // . is commonly used in OTel but disallowed in GCM label names,
+    // https://cloud.google.com/monitoring/api/ref_v3/rest/v3/LabelDescriptor#:~:text=Matches%20the%20following%20regular%20expression%3A
+    return key.replace('.', '_');
   }
 }