+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/admin/database/v1/DatabaseAdminClient.java
@@ -66,6 +66,8 @@ import com.google.spanner.admin.database.v1.GetDatabaseDdlRequest;
 import com.google.spanner.admin.database.v1.GetDatabaseDdlResponse;
 import com.google.spanner.admin.database.v1.GetDatabaseRequest;
 import com.google.spanner.admin.database.v1.InstanceName;
+import com.google.spanner.admin.database.v1.InternalUpdateGraphOperationRequest;
+import com.google.spanner.admin.database.v1.InternalUpdateGraphOperationResponse;
 import com.google.spanner.admin.database.v1.ListBackupOperationsRequest;
 import com.google.spanner.admin.database.v1.ListBackupOperationsResponse;
 import com.google.spanner.admin.database.v1.ListBackupSchedulesRequest;
@@ -637,6 +639,25 @@ import javax.annotation.Generated;
  *      <ul>
  *           <li><p> listBackupSchedulesPagedCallable()
  *           <li><p> listBackupSchedulesCallable()
+ *      </ul>
+ *       </td>
+ *    </tr>
+ *    <tr>
+ *      <td><p> InternalUpdateGraphOperation</td>
+ *      <td><p> This is an internal API called by Spanner Graph jobs. You should never need to call this API directly.</td>
+ *      <td>
+ *      <p>Request object method variants only take one parameter, a request object, which must be constructed before the call.</p>
+ *      <ul>
+ *           <li><p> internalUpdateGraphOperation(InternalUpdateGraphOperationRequest request)
+ *      </ul>
+ *      <p>"Flattened" method variants have converted the fields of the request object into function parameters to enable multiple ways to call the same method.</p>
+ *      <ul>
+ *           <li><p> internalUpdateGraphOperation(DatabaseName database, String operationId)
+ *           <li><p> internalUpdateGraphOperation(String database, String operationId)
+ *      </ul>
+ *      <p>Callable method variants take no parameters and return an immutable API callable object, which can be used to initiate calls to the service.</p>
+ *      <ul>
+ *           <li><p> internalUpdateGraphOperationCallable()
  *      </ul>
  *       </td>
  *    </tr>
@@ -5117,6 +5138,146 @@ public class DatabaseAdminClient implements BackgroundResource {
   public final UnaryCallable<ListBackupSchedulesRequest, ListBackupSchedulesResponse>
       listBackupSchedulesCallable() {
     return stub.listBackupSchedulesCallable();
+  }
+
+  // AUTO-GENERATED DOCUMENTATION AND METHOD.
+  /**
+   * This is an internal API called by Spanner Graph jobs. You should never need to call this API
+   * directly.
+   *
+   * <p>Sample code:
+   *
+   * <pre>{@code
+   * // This snippet has been automatically generated and should be regarded as a code template only.
+   * // It will require modifications to work:
+   * // - It may require correct/in-range values for request initialization.
+   * // - It may require specifying regional endpoints when creating the service client as shown in
+   * // https://cloud.google.com/java/docs/setup#configure_endpoints_for_the_client_library
+   * try (DatabaseAdminClient databaseAdminClient = DatabaseAdminClient.create()) {
+   *   DatabaseName database = DatabaseName.of("[PROJECT]", "[INSTANCE]", "[DATABASE]");
+   *   String operationId = "operationId129704162";
+   *   InternalUpdateGraphOperationResponse response =
+   *       databaseAdminClient.internalUpdateGraphOperation(database, operationId);
+   * }
+   * }</pre>
+   *
+   * @param database Internal field, do not use directly.
+   * @param operationId Internal field, do not use directly.
+   * @throws com.google.api.gax.rpc.ApiException if the remote call fails
+   */
+  public final InternalUpdateGraphOperationResponse internalUpdateGraphOperation(
+      DatabaseName database, String operationId) {
+    InternalUpdateGraphOperationRequest request =
+        InternalUpdateGraphOperationRequest.newBuilder()
+            .setDatabase(database == null ? null : database.toString())
+            .setOperationId(operationId)
+            .build();
+    return internalUpdateGraphOperation(request);
+  }
+
+  // AUTO-GENERATED DOCUMENTATION AND METHOD.
+  /**
+   * This is an internal API called by Spanner Graph jobs. You should never need to call this API
+   * directly.
+   *
+   * <p>Sample code:
+   *
+   * <pre>{@code
+   * // This snippet has been automatically generated and should be regarded as a code template only.
+   * // It will require modifications to work:
+   * // - It may require correct/in-range values for request initialization.
+   * // - It may require specifying regional endpoints when creating the service client as shown in
+   * // https://cloud.google.com/java/docs/setup#configure_endpoints_for_the_client_library
+   * try (DatabaseAdminClient databaseAdminClient = DatabaseAdminClient.create()) {
+   *   String database = DatabaseName.of("[PROJECT]", "[INSTANCE]", "[DATABASE]").toString();
+   *   String operationId = "operationId129704162";
+   *   InternalUpdateGraphOperationResponse response =
+   *       databaseAdminClient.internalUpdateGraphOperation(database, operationId);
+   * }
+   * }</pre>
+   *
+   * @param database Internal field, do not use directly.
+   * @param operationId Internal field, do not use directly.
+   * @throws com.google.api.gax.rpc.ApiException if the remote call fails
+   */
+  public final InternalUpdateGraphOperationResponse internalUpdateGraphOperation(
+      String database, String operationId) {
+    InternalUpdateGraphOperationRequest request =
+        InternalUpdateGraphOperationRequest.newBuilder()
+            .setDatabase(database)
+            .setOperationId(operationId)
+            .build();
+    return internalUpdateGraphOperation(request);
+  }
+
+  // AUTO-GENERATED DOCUMENTATION AND METHOD.
+  /**
+   * This is an internal API called by Spanner Graph jobs. You should never need to call this API
+   * directly.
+   *
+   * <p>Sample code:
+   *
+   * <pre>{@code
+   * // This snippet has been automatically generated and should be regarded as a code template only.
+   * // It will require modifications to work:
+   * // - It may require correct/in-range values for request initialization.
+   * // - It may require specifying regional endpoints when creating the service client as shown in
+   * // https://cloud.google.com/java/docs/setup#configure_endpoints_for_the_client_library
+   * try (DatabaseAdminClient databaseAdminClient = DatabaseAdminClient.create()) {
+   *   InternalUpdateGraphOperationRequest request =
+   *       InternalUpdateGraphOperationRequest.newBuilder()
+   *           .setDatabase(DatabaseName.of("[PROJECT]", "[INSTANCE]", "[DATABASE]").toString())
+   *           .setOperationId("operationId129704162")
+   *           .setVmIdentityToken("vmIdentityToken-417652124")
+   *           .setProgress(-1001078227)
+   *           .setStatus(Status.newBuilder().build())
+   *           .build();
+   *   InternalUpdateGraphOperationResponse response =
+   *       databaseAdminClient.internalUpdateGraphOperation(request);
+   * }
+   * }</pre>
+   *
+   * @param request The request object containing all of the parameters for the API call.
+   * @throws com.google.api.gax.rpc.ApiException if the remote call fails
+   */
+  public final InternalUpdateGraphOperationResponse internalUpdateGraphOperation(
+      InternalUpdateGraphOperationRequest request) {
+    return internalUpdateGraphOperationCallable().call(request);
+  }
+
+  // AUTO-GENERATED DOCUMENTATION AND METHOD.
+  /**
+   * This is an internal API called by Spanner Graph jobs. You should never need to call this API
+   * directly.
+   *
+   * <p>Sample code:
+   *
+   * <pre>{@code
+   * // This snippet has been automatically generated and should be regarded as a code template only.
+   * // It will require modifications to work:
+   * // - It may require correct/in-range values for request initialization.
+   * // - It may require specifying regional endpoints when creating the service client as shown in
+   * // https://cloud.google.com/java/docs/setup#configure_endpoints_for_the_client_library
+   * try (DatabaseAdminClient databaseAdminClient = DatabaseAdminClient.create()) {
+   *   InternalUpdateGraphOperationRequest request =
+   *       InternalUpdateGraphOperationRequest.newBuilder()
+   *           .setDatabase(DatabaseName.of("[PROJECT]", "[INSTANCE]", "[DATABASE]").toString())
+   *           .setOperationId("operationId129704162")
+   *           .setVmIdentityToken("vmIdentityToken-417652124")
+   *           .setProgress(-1001078227)
+   *           .setStatus(Status.newBuilder().build())
+   *           .build();
+   *   ApiFuture<InternalUpdateGraphOperationResponse> future =
+   *       databaseAdminClient.internalUpdateGraphOperationCallable().futureCall(request);
+   *   // Do something.
+   *   InternalUpdateGraphOperationResponse response = future.get();
+   * }
+   * }</pre>
+   */
+  public final UnaryCallable<
+          InternalUpdateGraphOperationRequest, InternalUpdateGraphOperationResponse>
+      internalUpdateGraphOperationCallable() {
+    return stub.internalUpdateGraphOperationCallable();
   }
 
   @Override