+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/connection/SpannerPool.java
@@ -164,6 +164,7 @@ public class SpannerPool {
     private final String clientCertificate;
     private final String clientCertificateKey;
     private final boolean isExperimentalHost;
+    private final Boolean enableDirectAccess;
 
     @VisibleForTesting
     static SpannerPoolKey of(ConnectionOptions options) {
@@ -198,6 +199,7 @@ public class SpannerPool {
       this.clientCertificate = options.getClientCertificate();
       this.clientCertificateKey = options.getClientCertificateKey();
       this.isExperimentalHost = options.isExperimentalHost();
+      this.enableDirectAccess = options.isEnableDirectAccess();
     }
 
     @Override
@@ -223,7 +225,8 @@ public class SpannerPool {
           && Objects.equals(this.enableEndToEndTracing, other.enableEndToEndTracing)
           && Objects.equals(this.clientCertificate, other.clientCertificate)
           && Objects.equals(this.clientCertificateKey, other.clientCertificateKey)
-          && Objects.equals(this.isExperimentalHost, other.isExperimentalHost);
+          && Objects.equals(this.isExperimentalHost, other.isExperimentalHost)
+          && Objects.equals(this.enableDirectAccess, other.enableDirectAccess);
     }
 
     @Override
@@ -245,7 +248,8 @@ public class SpannerPool {
           this.enableEndToEndTracing,
           this.clientCertificate,
           this.clientCertificateKey,
-          this.isExperimentalHost);
+          this.isExperimentalHost,
+          this.enableDirectAccess);
     }
   }
 
@@ -411,6 +415,9 @@ public class SpannerPool {
     }
     if (key.isExperimentalHost) {
       builder.setExperimentalHost(key.host);
+    }
+    if (key.enableDirectAccess != null) {
+      builder.setEnableDirectAccess(key.enableDirectAccess);
     }
     if (options.getConfigurator() != null) {
       options.getConfigurator().configure(builder);